remove SUBTREELoaderBase.js
